Rename HY_485__CSD4700_HW6.ipynb to Analyzing Sunspot Activity: Stationarity and Time Series Patterns.ipynb
In this asignement we begin by exploring the sunspot dataset, a well-known resource in time series analysis, focusing specifically on the monthly sunspot activity. After loading the data, a visual representation of the sunspot numbers over time is created, allowing us to observe any apparent trends or patterns. To ascertain the statistical nature of this time series, particularly its stationarity, we utilize the Augmented Dickey-Fuller (ADF) test. The results of this test, including the statistical value, the probability value, and critical thresholds, are then presented. These results are essential for understanding the underlying characteristics of the sunspot data and for determining the appropriate modeling techniques.